Hide MPI calls behind `HAVE_MPI`
Hide MPI calls behind preprocessor macro `HAVE_MPI` in `ezmpi_recv_m`
and `ezmpi_send_m`. Thus a non-MPI program can link with EZMPI or a
library requiring EZMPI.

In `ezmpi_recv_m`, we hide all the content of the module because the
argument buf of `ezmpi_recv` is `intent(out)` so `ezmpi_recv` should
not be called in a non-MPI program.